Getting rid of device containers.
Device containers were a thoughtful way of separating device access from
the implementation of the containers. But typedefs are frankly not ideal
for this purpose, since the user must look up the typedef in order to
use the object. Moreover, the syntax of using device containers was
clunky and it is now easier for clients to use the device factory
without the verbose syntax.

Also switching device containers from lists to vectors since these
containers should be unmodifiable (the devices can be modified but
container itself should be const - added this as well) and not only are
vectors a better choice in general, this means there should be no
occasion where a list might be superior.