Reconnect after device changes, add dieHorribly
Reconnect is no longer a valid script command. Instead, that
happens every time you enable or disable a device. Note that that
goes through the normal polite quit protocol with DGI, so some
delay is expected.

Also adds support for dieHorribly, so we can simulate controller
glitches and network failures. Right now the controller "recovers"
instantaneously... I'll upgrade that in the next commit.