Add fake DGI for testing the fake controller
Seems easiest to be sure the controller script is working before
testing with the real DGI, no?

The fake DGI isn't QUITE ready yet, but it almost is.

Then we just have to test the fake DGI for testing the fake
controller for testing the real DGI, and we're good on the fake
controller side of things.