Handle DGI responses to custom packets
Incidentally this also makes the controller hang. Possibly the DGI
isn't sending responses right away? With the previous commit, the
DGI would send an empty BadRequest after the controller attempted
to PoliteDisconnect after sending a custom packet to the adapter.
So not really sure what's going on, but you may want to try
rolling back this commit to see the controller not hang after
sending a custom packet.

Also, add UnexpectedError9, send a Golbat to AdapterFactory. This
currently crashes DGI.

Relevant to Issue #155