Don't throw exceptions out of CRtdsAdapter::Run
Since this function is the top level of its thread, throwing an
exception from here leads to an abort (and a core dump if your
system is configured to make those). Shut down cleanly instead.

Note that it's correct for send/receive to be fatal here, as we
assume a completely reliable connection for CRtdsAdapter. In
practice this will only occur if the DGI is misconfigured or the
fpga (or pscad interface) stops listening.