Revert "simserv: drop rtdsadapter nan values before sending to pscad"
This reverts commit a2f7f8e0b684bcf11677f7b6ba81d9d26ab400a7.

For starters, this shouldn't have been pushed without a pull request.
That process would have caught the breakage.

The DGI in master sends 0 as device commands even if it doesn't know
anything about the state of its devices, in contrast to the PnP DGI
which uses NaN in this situation. Accordingly we need to filter not
NaN commands but 0 commands. No this doesn't make sense, but it works
(except for the portion of the convergence where you actually do want
the state set to 0).

This commit can safely be reverted in master because we have already
pushed a different, functionally-equivalent commit to PnP, so this
will not undo any PnP code when merged into PnP.

Note that this means PnP and master no longer share the same simserv,
so we just have to be careful to recompile that when switching
branches.