Stop: remove some code that's not doing anything
Deleting the workload to allow the io_service to run to completion, then
immediately stopping it before giving that a chance to occur, was
pointless.
Let's just do the stop (which is always safe) and be done with it.  This
preserves the current behavior, except in the unlikely case that the
service
could have run to completion between the delete and the stop.

The disadvantage of this is that we don't get to finish all of our async
operations as desired.  But again, that's current behavior as well, and
I don't see an easy way around it.

This is worth some further discussion.