CBroker: assert that the round is nonzero
The Clang static analyzer found a potential division by zero in
CBroker::ChangePhase(). The bug occurs when (a) there is at least one
module registered, and (b) every registered module has a phase duration
of zero. Obviously this is not a practical concern, but it would be good
to add an assertion to surpress the false positive. (It's also
independently not a bad idea.)