Remove custom message expiration times
Currently, we support the ability to set custom expiration times for
each message. This is used only by group management, and there is nasty
hardcoded support to allow group management (and only group management)
to access the expiration time of the messages it receives, for the
purpose of setting custom expiration times.

It turns out that group management does not actually need this ability:
the custom expiration times are just a hint to the IProtocol. So let's
remove this.